requirements.txt
# Core dependencies
openai>=1.10.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0

# PDF processing
PyPDF2>=3.0.0
nltk>=3.8.1

# For caching
diskcache>=5.6.0

# For retry logic
backoff>=2.2.1

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

# Pinecone for vector database (to be implemented)
pinecone-client>=2.2.4 